fix: change location id data type for Long to int

+++ src/main/java/com/company/repository/LocationRepository.java
@@ -4,7 +4,7 @@ import com.company.model.entity.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface LocationRepository extends JpaRepository<Location, Long> {
+public interface LocationRepository extends JpaRepository<Location, Integer> {
 
     Location findByCountryAndCityAndState(String country, String city, String state);
